Add method for counting calories of our recipe(with double value)

+++ src/main/java/com/exigen/common/service/RecipeService.java
@@ -71,5 +71,5 @@ public interface RecipeService {
      * @throws NullPointerException(when recipe's id has no results in the database)
      */
 
-    float CaloriesCalculation(List<Integer> recipesId);
+    double CaloriesCalculation(List<Integer> recipesId);
 }
